requirements.txt
streamlit>=1.28.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
langchain-core>=0.2.0
pillow>=10.0.0  # For image processing